+++ docs/changes.rst
@@ -149,6 +149,18 @@ backends:
   This makes it possible to support lookup of artist or album URIs which then
   can expand to a list of tracks.
 
+- Remove :attr:`mopidy.core.PlaybackController.track_at_previous`. Use
+  :attr:`mopidy.core.PlaybackController.tl_track_at_previous` instead.
+
+- Remove :attr:`mopidy.core.PlaybackController.track_at_next`. Use
+  :attr:`mopidy.core.PlaybackController.tl_track_at_next` instead.
+
+- Remove :attr:`mopidy.core.PlaybackController.track_at_eot`. Use
+  :attr:`mopidy.core.PlaybackController.tl_track_at_eot` instead.
+
+- Remove :attr:`mopidy.core.PlaybackController.current_tlid`. Use
+  :attr:`mopidy.core.PlaybackController.current_tl_track` instead.
+
 - Added support for connecting to the Spotify service through an HTTP or SOCKS
   proxy, which is supported by pyspotify >= 1.9.